+++ lib/dradis/plugins/openvas/gem_version.rb
@@ -10,7 +10,7 @@ module Dradis
         MAJOR = 3
         MINOR = 18
         TINY = 0
-        PRE = nil
+        PRE = 'rc1'
 
         STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
       end