+++ spec/spec_helper.rb
@@ -1,35 +1,12 @@
-ENV["RAILS_ENV"] ||= 'test'
-require File.expand_path("../../../../../config/environment", __FILE__)
-require 'rspec/rails'
+require 'rubygems'
+require 'bundler/setup'
+require 'nokogiri'
 
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-require 'support/fixture_loader'
+require 'combustion'
+
+Combustion.initialize!
 
 RSpec.configure do |config|
-  # CLI niceties
-  config.order = :random
-
-  # Filter which specs to run
-  config.treat_symbols_as_metadata_keys_with_true_values = true
-  config.filter_run :focus => true
-  config.run_all_when_everything_filtered = true
-
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = false
-
-  config.before(:suite) do
-    DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.clean_with(:truncation)
-  end
-
-  config.before(:each) do
-    DatabaseCleaner.start
-  end
-
-  config.after(:each) do
-    DatabaseCleaner.clean
-  end
 end
+
+require 'support/fixture_loader'